Refactoring to better named components and using tabs for viewing org details
